+++ sdk/metric/instrument.go
@@ -241,7 +241,7 @@ func newInt64Observable(scope instrumentation.Scope, kind InstrumentKind, name, 
 }
 
 type observable[N int64 | float64] struct {
-	instrument.Asynchronous
+	instrument.Observable
 	observablID[N]
 
 	aggregators []internal.Aggregator[N]